+++ servo/components/layout/parallel.rs
@@ -109,7 +109,8 @@ pub trait ParallelPreorderDomTraversal : PreorderDomTraversal {
         {
             let mut layout_data_ref = node.mutate_layout_data();
             let layout_data = layout_data_ref.as_mut().expect("no layout data");
-            layout_data.data.parallel.children_count.store(child_count as int, Ordering::Relaxed);
+            layout_data.data.parallel.children_count.store(child_count as isize,
+                                                           Ordering::Relaxed);
         }
 
         
@@ -231,7 +232,8 @@ trait ParallelPostorderFlowTraversal : PostorderFlowTraversal {
                 let base = flow::mut_base(&mut **flow);
 
                 
-                base.parallel.children_count.store(base.children.len() as int, Ordering::Relaxed);
+                base.parallel.children_count.store(base.children.len() as isize,
+                                                   Ordering::Relaxed);
 
                 
                 let unsafe_parent = base.parallel.parent;
